+++ backend/canisters/community/api/src/updates/join_video_call.rs
@@ -1,9 +1,10 @@
+use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
 use types::{ChannelId, MessageId, UnitResult, VideoCallPresence};
 
 #[ts_export(community, join_video_call)]
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub channel_id: ChannelId,
     pub message_id: MessageId,