+++ backend/canisters/community/api/src/updates/start_video_call_v2.rs
@@ -1,7 +1,8 @@
+use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{ChannelId, MessageId, Milliseconds, UnitResult, UserId, VideoCallType};
 
-#[derive(Serialize, Deserialize, Debug)]
+#[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
     pub channel_id: ChannelId,
     pub message_id: MessageId,