+++ backend/canisters/group/api/src/queries/search_messages.rs
@@ -1,4 +1,3 @@
-use candid::CandidType;
 use oc_error_codes::OCError;
 use serde::{Deserialize, Serialize};
 use std::collections::HashSet;
@@ -6,7 +5,7 @@ use ts_export::ts_export;
 use types::{MessageMatch, UserId};
 
 #[ts_export(group, search_messages)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub search_term: String,
     pub max_results: u8,
@@ -14,14 +13,14 @@ pub struct Args {
 }
 
 #[ts_export(group, search_messages)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub enum Response {
     Success(SuccessResult),
     Error(OCError),
 }
 
 #[ts_export(group, search_messages)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct SuccessResult {
     pub matches: Vec<MessageMatch>,
 }