+++ backend/canisters/group/api/src/updates/c2c_bot_invite_users.rs
@@ -1,4 +1,4 @@
-use candid::Principal;
+use ic_principal::Principal;
 use serde::{Deserialize, Serialize};
 use types::{BotInitiator, UnitResult, UserId};