+++ backend/canisters/group/api/src/updates/c2c_tip_message.rs
@@ -1,8 +1,7 @@
-use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use types::{CanisterId, MessageId, MessageIndex, UnitResult, UserId};
 
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub recipient: UserId,
     pub thread_root_message_index: Option<MessageIndex>,