+++ backend/canisters/group/api/src/updates/edit_message_v2.rs
@@ -1,10 +1,9 @@
-use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
 use types::{MessageContentInitial, MessageId, MessageIndex, UnitResult};
 
 #[ts_export(group, edit_message)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub thread_root_message_index: Option<MessageIndex>,
     pub message_id: MessageId,