+++ backend/canisters/group/api/src/updates/toggle_mute_notifications.rs
@@ -1,10 +1,9 @@
-use candid::CandidType;
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
 use types::UnitResult;
 
 #[ts_export(group, toggle_mute_notifications)]
-#[derive(CandidType, Serialize, Deserialize, Debug)]
+#[derive(Serialize, Deserialize, Debug)]
 pub struct Args {
     pub mute: bool,
 }