+++ backend/canisters/local_user_index/CHANGELOG.md
@@ -11,6 +11,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Push full chat events to subscribed bots ([#8264](https://github.com/open-chat-labs/open-chat/pull/8264))
 - Use MessagePack to serialize upgrade args for Users/Group/Communities ([#8269](https://github.com/open-chat-labs/open-chat/pull/8269))
 
+### Fixed
+
+- Trying to fix fcm_data issue ([#8272](https://github.com/open-chat-labs/open-chat/pull/8272))
+
 ## [[2.0.1809](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1809-local_user_index)] - 2025-06-27
 
 ### Changed